+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added class and driver nodes instanciation
 - Added REPL custom widget in control view
 - Added generic widgets for control view nodes
 - Added connection state management for control view nodes